requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
datasets==2.19.0
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.109.1
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.3.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
limits==4.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy==1.15.0
mypy-extensions==1.0.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.26.4
openai==0.28.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
propcache==0.2.1
psutil==7.0.0
psycopg2==2.9.10
pyarrow==19.0.1
pyarrow-hotfix==0.6
pydantic==2.10.4
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.0
tqdm==4.67.1
transformers==4.49.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.27.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
